+++ spec/truffle/interop/foreign_inspect_to_s_spec.rb
@@ -39,8 +39,7 @@ describe "#inspect and #to_s on a foreign" do
         hash = {a: 1, b: 2, c: 3}
         foreign = Truffle::Interop.to_java_map(hash)
 
-        hash.to_s.should == '{:a=>1, :b=>2, :c=>3}' # for comparison
-        foreign.inspect.should =~ /\A#<Polyglot::ForeignHash\[Java\] java\.util\.HashMap:0x\h+ {"a"=>1, "b"=>2, "c"=>3}>\z/
+        foreign.inspect.should =~ /\A#<Polyglot::ForeignHash\[Java\] java\.util\.HashMap:0x\h+ {"a" => 1, "b" => 2, "c" => 3}>\z/
         foreign.to_s.should == "#<Polyglot::ForeignHash[Java] {a=1, b=2, c=3}>"
       end
     end
@@ -160,8 +159,7 @@ describe "#inspect and #to_s on a foreign" do
   describe "hash" do
     it "gives a similar representation to Ruby" do
       foreign = Truffle::Debug.foreign_hash
-      { a: 1, b: 2 }.inspect.should == "{:a=>1, :b=>2}"
-      foreign.inspect.should =~ /\A#<Polyglot::ForeignHash:0x\h+ {:a=>1, :b=>2}>\z/
+      foreign.inspect.should =~ /\A#<Polyglot::ForeignHash:0x\h+ {:a => 1, :b => 2}>\z/
       foreign.to_s.should == "#<Polyglot::ForeignHash [foreign hash]>"
     end
   end